+++ packages/eslint-plugin/src/__tests__/fixtures/require-cached-repo-provider/uncached/repositories/image.repository.ts
@@ -1,0 +1,3 @@
+export class ProductImageRepository extends defineAppRepo({
+  model: 'productImage',
+}) {}